Move TreeNode extensions to tests package
