+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2020-09-17 12:16","gitSha":"81e47c4f4"}
+{"buildTime":"2020-09-21 12:49","gitSha":"b3d866a39"}
